+++ data/src/main/java/dk/itu/data/utils/RoutingUtils.java
@@ -6,10 +6,10 @@ import java.util.HashMap;
 import java.util.Map;
 
 public class RoutingUtils {
-    private static final double EARTH_RADIUS_KM = 6371;
+    private static final float EARTH_RADIUS_KM = 6371f;
 
     // This is based on the haversine formula to calculate distance between two points while accounting for earths curvature
-    public static Double distanceMeters(double lat1, double lon1, double lat2, double lon2){
+    public static Float distanceMeters(float lat1, float lon1, float lat2, float lon2){
         double lat1Rad = Math.toRadians(lat1);
         double lon1Rad = Math.toRadians(lon1);
         double lat2Rad = Math.toRadians(lat2);
@@ -22,7 +22,7 @@ public class RoutingUtils {
 
         double c = 2 * Math.atan2(Math.sqrt(a), Math.sqrt(1 - a));
 
-        return EARTH_RADIUS_KM * c * 1000;
+        return (float) (EARTH_RADIUS_KM * c * 1000);
     }
 
     public static float distanceMetersFloat(double lat1, double lon1, double lat2, double lon2){
